Restructured geometry shapes
