+++ core/client_side/game_mechanics/logic/game_components/ball.js
@@ -1,13 +1,13 @@
 
 import * as math from '../../../../_lib/math';
-import * as shapes from '../geometry_shapes';
+import {Circle} from '../geometry_shapes/circle';
 import {SolidBody} from '../solid_body';
 
 
 export class Ball extends SolidBody {
     constructor(radius) {
         super();
-        this._circle = new shapes.Circle(radius);
+        this._circle = new Circle(radius);
     }
 
     get radius() {
