+++ core/client_side/game_mechanics/logic/geometry_shapes/circle.js
@@ -1,0 +1,10 @@
+
+export class Circle {
+    constructor(radius) {
+        this._radius = radius;
+    }
+
+    get radius() {
+        return this._radius;
+    }
+}
