+++ core/client_side/game_mechanics/logic/geometry_shapes/triangle.js
@@ -1,14 +1,4 @@
 
-export class Circle {
-    constructor(radius) {
-        this._radius = radius;
-    }
-
-    get radius() {
-        return this._radius;
-    }
-}
-
 
 export class Triangle {
     constructor(height, sectorAngle) {
